make some update

+++ database/migrations/2021_02_17_135050_create_products_models_table.php
@@ -16,7 +16,7 @@ class CreateProductsModelsTable extends Migration
         Schema::create('products_models', function (Blueprint $table) {
             $table->id();
             $table->string('name');
-            $table->double('price', 8, 2);
+            $table->decimal('price', 10, 2);
             $table->timestamps();
         });
     }
